backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.10.1
python-multipart==0.0.20
psycopg2-binary==2.9.10
pywebpush==2.0.3
boto3==1.40.69
slowapi==0.1.9
